refactor: promote workflow_* tools from registry to core
The 8 ask-built-in workflow tools (workflow_list, workflow_get,
workflow_create, workflow_edit, workflow_delete, workflow_copy,
workflow_run, clear_plans) now ride the wire as core tools instead
of living in the deferred registry reached via search_tools +
invoke_tool. The model sees them directly in its tool definitions
the same way it sees read, bash, todos, and end_turn.

Why: the two-stage workflow guard in agent_tools_todos.go steers
the model to call workflow_list (stage 1) and workflow_run
(stage 2) as preconditions for multi-step work. With the tools in
the registry, every such call cost an extra search_tools round-trip
plus invoke_tool indirection. Moving them to the wire removes that
overhead and makes the guard's required first actions a single
direct call.

The disarm hooks (env.markWorkflowsChecked,
env.markWorkflowRunDispatched) moved from the registry's
agentInvokeTool.Run into the workflow_list and workflow_run tool
closures themselves, so the guard clears whether the tool is
invoked via the wire or via some future path. The
agentSearchToolsDescription and the "is a core tool" error path
in invoke_tool keep the rest of the surface consistent.

agentBridgeTools in agent_tools_bridge.go was renamed to
agentLinearTools and now returns only the 12 linear_* tools; the
shared nativeBridgeTool / bridgeAgentTool / mcpResultText helpers
are reused by the new agent_tools_workflow.go.

Tests: 5 workflow tests moved from agent_tools_bridge_test.go into
a new agent_tools_workflow_test.go and re-sourced from
agentWorkflowTools. Added TestAgentWorkflowTools_DisarmHooksFire
that exercises the real workflow tool cores. Updated
agent_tools_registry_test.go fixtures (workflow_* and clear_plans
are no longer registry tools), agent_tools_bridge_test.go
(linear_* coverage only), agent_tools_test.go (workflow guard
disarms via the real tool cores, and
TestCoreTools_RequireDescriptionPhrase now includes the workflow
tools and verifies the workflow_create / workflow_edit
"description-is-payload" exception), and mcp_workflows_test.go
(EditDecodesStepsJSON iterates agentWorkflowTools).

Docs: CLAUDE.md / AGENTS.md file table and the "Tool registry vs
core tools" callout now list the workflow tools alongside
web_search as documented core exceptions, with the workflow guard
cited as the reason. The guard notices in agent_tools_todos.go
and the multi-step precondition in provider.go::askSteeringPrompt
now steer the model to call the tools directly.
