+++ agent_tools_bridge.go
@@ -11,14 +11,22 @@ import (
 )
 
 // agent_tools_bridge.go gives the in-process agent sessions native
-// twins of every loopback-bridge tool (linear_*, workflow_*) so they
-// no longer attach the HTTP bridge at all. The bridge handlers and
-// these natives share the same cwd-parameterized core functions
-// (mcp_linear.go / mcp_workflows.go), so behavior — gating, errors,
-// output shapes — is identical on both paths. These tools live in
-// the deferred tool registry (agent_tools_registry.go): they are
-// never part of the wire tool definitions and are reached through
-// search_tools + invoke_tool.
+// twins of the linear_* bridge tools so they no longer attach the HTTP
+// bridge at all. The bridge handlers and these natives share the same
+// cwd-parameterized core functions in mcp_linear.go, so behavior —
+// gating, errors, output shapes — is identical on both paths. These
+// tools live in the deferred tool registry (agent_tools_registry.go):
+// they are never part of the wire tool definitions and are reached
+// through search_tools + invoke_tool.
+//
+// The 8 ask-built-in workflow_* tools (and clear_plans) used to live
+// here too but were promoted to the core wire toolset — see
+// agent_tools_workflow.go. The promotion is a deliberate, documented
+// core exception (the two-stage workflow guard forces the model to
+// call workflow_list and workflow_run as a precondition for multi-step
+// work; an extra search_tools round-trip per guard interaction was
+// pure overhead). nativeBridgeTool, bridgeAgentTool, and
+// mcpResultText are still reused by agent_tools_workflow.go.
 
 // nativeBridgeTool adapts one bridge core to fantasy's AgentTool. The
 // input schema is generated by the same jsonschema machinery the MCP
@@ -146,12 +154,12 @@ func mcpResultText(res *mcp.CallToolResult) string {
 	return b.String()
 }
 
-// agentBridgeTools builds the full native tool set for one session:
-// the 12 linear_* tools and the 8 workflow_* tools (CRUD + copy +
-// run + clear_plans). ask_user_question and end_turn already have
-// natives in agent_tools_ask.go. Tool names match what the bridge
-// exposes so prompts and saved workflows keep working unchanged.
-func agentBridgeTools(env *agentToolEnv) []fantasy.AgentTool {
+// agentLinearTools builds the 12 linear_* native twins for one
+// session. The ask-built-in workflow tools used to live here too but
+// were promoted to the core wire toolset in agent_tools_workflow.go.
+// Tool names match what the bridge exposes so prompts and saved
+// workflows keep working unchanged.
+func agentLinearTools(env *agentToolEnv) []fantasy.AgentTool {
 	cwd := func() string { return env.cwd }
 	return []fantasy.AgentTool{
 		nativeBridgeTool("linear_list_issues", linearListToolDescription,
@@ -201,39 +209,6 @@ func agentBridgeTools(env *agentToolEnv) []fantasy.AgentTool {
 		nativeBridgeTool("linear_list_cycles", linearListCyclesToolDescription,
 			func(ctx context.Context, in linearListCyclesInput) (*mcp.CallToolResult, linearListCyclesOutput, error) {
 				return linearListCyclesCore(ctx, cwd(), in)
-			}),
-
-		nativeBridgeTool("workflow_list", workflowListToolDescription,
-			func(_ context.Context, in workflowListInput) (*mcp.CallToolResult, workflowListOutput, error) {
-				return workflowListCore(cwd(), in)
-			}),
-		nativeBridgeTool("workflow_get", workflowGetToolDescription,
-			func(_ context.Context, in workflowGetInput) (*mcp.CallToolResult, workflowGetOutput, error) {
-				return workflowGetCore(cwd(), in)
-			}),
-		nativeBridgeTool("workflow_create", workflowCreateToolDescription,
-			func(_ context.Context, in workflowCreateInput) (*mcp.CallToolResult, workflowCreateOutput, error) {
-				return workflowCreateCore(cwd(), in)
-			}),
-		nativeBridgeTool("workflow_edit", workflowEditToolDescription,
-			func(_ context.Context, in workflowEditInput) (*mcp.CallToolResult, workflowEditOutput, error) {
-				return workflowEditCore(cwd(), in)
-			}),
-		nativeBridgeTool("workflow_delete", workflowDeleteToolDescription,
-			func(_ context.Context, in workflowDeleteInput) (*mcp.CallToolResult, workflowDeleteOutput, error) {
-				return workflowDeleteCore(cwd(), in)
-			}),
-		nativeBridgeTool("workflow_copy", workflowCopyToolDescription,
-			func(_ context.Context, in workflowCopyInput) (*mcp.CallToolResult, workflowCopyOutput, error) {
-				return workflowCopyCore(cwd(), in)
-			}),
-		nativeBridgeTool("workflow_run", workflowRunToolDescription,
-			func(_ context.Context, in workflowRunInput) (*mcp.CallToolResult, workflowRunOutput, error) {
-				return workflowRunCore(cwd(), env.tabID, in)
-			}),
-		nativeBridgeTool("clear_plans", clearPlansToolDescription,
-			func(_ context.Context, in clearPlansInput) (*mcp.CallToolResult, clearPlansOutput, error) {
-				return clearPlansCore(cwd(), in)
 			}),
 	}
 }
